Issue-3499: Move volume module to compose/core package
Partially implements #3499
